requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio[asyncio_client]>=5.10.0
httpx>=0.27.0
